moved facility into models
