+++ Density/app/src/main/java/org/cornelldti/density/density/BaseActivity.kt
@@ -19,6 +19,7 @@ import org.json.JSONException
 
 import java.text.SimpleDateFormat
 import androidx.appcompat.app.AppCompatActivity
+import org.cornelldti.density.density.models.Facility
 import java.util.Calendar
 import java.util.Locale
