+++ Density/app/src/main/java/org/cornelldti/density/density/FacilityPage.kt
@@ -27,6 +27,7 @@ import org.json.JSONException
 import java.text.SimpleDateFormat
 
 import androidx.core.content.ContextCompat
+import org.cornelldti.density.density.models.Facility
 import java.util.Calendar
 import java.util.Locale
 import java.util.Date
