+++ Density/app/src/main/java/org/cornelldti/density/density/MainActivity.kt
@@ -28,7 +28,7 @@ import androidx.appcompat.widget.SearchView
 import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import org.cornelldti.density.density.models.Facility
 import kotlin.math.absoluteValue
 
 class MainActivity : BaseActivity() {
